pyproject.toml
[project]
name = "wikimind"
version = "0.1.0"
description = "A personal wiki builder powered by LLMs"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pymupdf4llm>=0.0.17",
    "trafilatura>=2.0",
    "python-frontmatter>=1.0",
    "rank-bm25>=0.2.2",
    "flask>=3.0",
]

[project.scripts]
wikimind = "wikimind.cli:cli"

[tool.setuptools]
packages = ["wikimind"]
package-dir = {"wikimind" = "."}

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
